core: move Database into its own database.rs module
lib.rs had grown to ~5000 lines, most of it the Database struct, its
open/registry machinery, and DatabaseCatalog. Move all of that into a
new core/database.rs module so lib.rs is back to being the crate root:
module declarations, re-exports, and small shared types (CDC info,
SyncMode, TempStore, QueryRunner, database-id constants).

Pure code motion: every type and function moved verbatim, with lib.rs
re-exporting the same public names so the crate API is unchanged.
multiprocess_tests.rs now imports DatabaseFile and atomic Ordering
directly since it no longer gets them via lib.rs glob.

Tests: cargo test -p turso_core, cargo clippy --workspace --all-features --all-targets
