+++ core/multiprocess_tests.rs
@@ -1,4 +1,6 @@
 use super::*;
+use crate::storage::database::DatabaseFile;
+use crate::sync::atomic::Ordering;
 use std::process::Command;
 use std::time::Duration;
